requirements.txt
numpy==1.26.4
Pillow
onnxruntime==1.18